fix copied file
